+++ src/components/form/RadioField.tsx
@@ -9,6 +9,7 @@ import { FormFieldError } from "./typings";
 import useDisabled from "./useDisabled";
 
 export type RadioFieldValue = string | number;
+
 type RadioFieldOption = {
   label: string;
   value: RadioFieldValue;