+++ src/components/form/select/SelectFieldCaret.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import classnames from "../../../classnames";
 
-const SelectFieldCaret: React.FC<{ open: boolean }> = ({ open }) => (
+const SelectFieldCaret: React.FC<{ open: boolean }> = React.memo(({ open }) => (
   <div className={classnames("chq-ffd--sl--caret", { "chq-ffd--sl--caret-flip": open })} />
-);
+));
 
 export default SelectFieldCaret;